requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
SQLAlchemy>=1.4.0
asyncpg>=0.27.0
python-dotenv>=1.0.0
httpx>=0.24.0